requirements.txt
python-novaclient>=2.0
